added support for more data

+++ src/instagram_graph_api/data/business_account.cr
@@ -9,7 +9,7 @@ module InstagramGraphApi
       property username : String
       property ig_id : Int64
 
-      property followers_count : Int64
+      property followers_count : Int32
 
       property profile_picture_url : String
